# docs_2/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WhatsApp API MultiDevice",
    "version": "6.9.0",
    "description": "This API is used for sending whatsapp via API"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development Server"
    }
  ],
...
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/api/devices": {
      "get": {
        "operationId": "listDevices",
        "tags": [
          "Device Management"
        ],
        "summary": "List registered devices",
        "description": "Returns list of all registered WhatsApp devices",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status",
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "registered",
                "error",
                "stopped"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 25,
              "maximum": 100
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Device list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "devices": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "number",
                            "example": 1
                          },
                          "deviceHash": {
                            "type": "string",
                            "example": "a1b2c3d4e5f67890"
                          },
                          "status": {
                            "type": "string",
                            "example": "active"
                          },
                          "webhookUrl": {
                            "type": "string",
                            "format": "url",
                            "nullable": true
                          },
                          "webhookSecret": {
                            "type": "string",
                            "nullable": true
                          },
                          "statusWebhookUrl": {
                            "type": "string",
                            "format": "url",
                            "nullable": true
                          },
                          "statusWebhookSecret": {
                            "type": "string",
                            "nullable": true
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createDevice",
        "tags": [
          "Device Management"
        ],
        "summary": "Register new device",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "webhookUrl": {
                    "type": "string",
                    "format": "url",
                    "nullable": true,
                    "example": "https://meusite.com/webhook/messages"
                  },
                  "webhookSecret": {
                    "type": "string",
                    "nullable": true,
                    "example": "meu-secret-mensagens"
                  },
                  "statusWebhookUrl": {
                    "type": "string",
                    "format": "url",
                    "nullable": true,
                    "example": "https://meusite.com/webhook/status"
                  },
                  "statusWebhookSecret": {
                    "type": "string",
                    "nullable": true,
                    "example": "meu-secret-status"
                  },
                  "autoStart": {
                    "type": "boolean",
                    "default": true,
                    "example": true
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Device created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "deviceHash": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string"
                    },
                    "webhookUrl": {
                      "type": "string",
                      "format": "url",
                      "nullable": true
                    },
                    "webhookSecret": {
                      "type": "string",
                      "nullable": true
                    },
                    "statusWebhookUrl": {
                      "type": "string",
                      "format": "url",
                      "nullable": true
                    },
                    "statusWebhookSecret": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "updateDeviceInfo",
        "tags": [
          "Device Management"
        ],
        "summary": "Update device information",
        "description": "Updates a device's name, message webhooks, and status webhooks.",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "New Device Name"
                  },
                  "webhookUrl": {
                    "type": "string",
                    "format": "url",
                    "nullable": true,
                    "example": "https://meusite.com/webhook/messages"
                  },
                  "webhookSecret": {
                    "type": "string",
                    "nullable": true,
                    "example": "novo-secret-mensagens"
                  },
                  "statusWebhookUrl": {
                    "type": "string",
                    "format": "url",
                    "nullable": true,
                    "example": "https://meusite.com/webhook/status"
                  },
                  "statusWebhookSecret": {
                    "type": "string",
                    "nullable": true,
                    "example": "novo-secret-status"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Device updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Device updated successfully"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "deviceHash": {
                          "type": "string"
                        },
                        "name": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteDevice",
        "tags": [
          "Device Management"
        ],
        "summary": "Remove a device",
        "description": "Removes a device and its associated Docker container.",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          },
          {
            "name": "force",
            "in": "query",
            "required": false,
            "description": "Force removal even if the process cannot be stopped.",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device removed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Device removed successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Error removing device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInternalServer"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/info": {
      "get": {
        "operationId": "getDeviceInfo",
        "tags": [
          "Device Management"
        ],
        "summary": "Get information from a specific device",
        "description": "Returns details of a WhatsApp device, including status and QR Code (if available).",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device information",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "number",
                      "example": 1
                    },
                    "deviceHash": {
                      "type": "string",
                      "example": "a1b2c3d4e5f67890"
                    },
                    "name": {
                      "type": "string",
                      "example": "Main Service"
                    },
                    "status": {
                      "type": "string",
                      "example": "connected"
                    },
                    "processStatus": {
                      "type": "object",
                      "properties": {
                        "status": {
                          "type": "string",
                          "example": "running"
                        },
                        "containerId": {
                          "type": "string",
                          "example": "a1b2c3d4e5f6"
                        }
                      }
                    },
                    "lastSeen": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "createdAt": {
                      "type": "string",
                      "format": "date-time"
                    },
                    "webhookUrl": {
                      "type": "string",
                      "format": "url",
                      "nullable": true
                    },
                    "webhookSecret": {
                      "type": "string",
                      "nullable": true
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/restart": {
      "post": {
        "operationId": "restartDevice",
        "tags": [
          "Device Management"
        ],
        "summary": "Restart a device",
        "description": "Restarts the Docker container associated with a device.",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device restarted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Process restarted successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Error restarting device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInternalServer"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/start": {
      "post": {
        "operationId": "startDevice",
        "tags": [
          "Device Management"
        ],
        "summary": "Launch a device",
        "description": "Starts the Docker container associated with a device.",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Process started successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Error starting device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInternalServer"
                }
              }
            }
          }
        }
      }
    },
    "/api/devices/stop": {
      "post": {
        "operationId": "stopDevice",
        "tags": [
          "Device Management"
        ],
        "summary": "Stop a device",
        "description": "For the Docker container associated with a device.",
        "security": [
          {
            "basicAuth": []
          }
        ],
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Device stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "message": {
                      "type": "string",
                      "example": "Process stopped successfully"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Device not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBadRequest"
                }
              }
            }
          },
          "500": {
            "description": "Error stopping device",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInternalServer"
                }
              }
            }
          }
        }
      }
    },
    "/api/app/devices": {
      "get": {
        "operationId": "appDevices",
        "tags": [
          "app"
        ],
        "summary": "Get list connected devices",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorInternalServer"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "x-instance-id",
            "in": "header",
            "required": true,
            "description": "The deviceHash of the instance (e.g. a1b2c3d4e5f67890)",
            "schema": {
              "type": "string",
              "example": "a1b2c3d4e5f67890"
            }
          }
        ],
        "security": [
          {
            "basicAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    },
    "schemas": {
...
      "DeviceResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "SUCCESS"
          },
          "message": {
            "type": "string",
            "example": "Fetch device success"
          },
          "results": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string",
                  "example": "Aldino Kemal"
                },
                "device": {
                  "type": "string",
                  "example": "628960561XXX.0:64@s.whatsapp.net"
                }
              }
            }
          }
        }
      },
...
      "ErrorInternalServer": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INTERNAL_SERVER_ERROR",
            "description": "SYSTEM_CODE_ERROR"
          },
          "message": {
            "type": "string",
            "example": "you are not loggin",
            "description": "Detail error message"
          },
          "results": {
            "type": "object",
            "example": null,
            "description": "additional data"
          }
        }
      },
      "ErrorBadRequest": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": 400,
            "description": "HTTP Status Code"
          },
          "message": {
            "type": "string",
            "example": "field cannot be blank",
            "description": "Detail error message"
          },
          "results": {
            "type": "object",
            "example": null,
            "description": "additional data"
          }
        }
      },
...
    }
  }
}